requirements.txt
hydra-core==1.3.2
loguru==0.7.2
matplotlib==3.7.5
mmdet==3.3.0
numpy==1.23.0
opencv_python==4.10.0.84
smplx==0.1.28
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
yacs==0.1.8
